Better refactor
